requirements.txt
Pillow==10.1.0
PyPDF2==3.0.1
tkinter-tooltip==2.1.0
customtkinter==5.2.0
